[svn-r7606] Purpose: Bug fix
Description: Type of the third argument to the h5pset(get)_cache_f subroutines
             was incorrect (INTEGER instead of INTEGER(SIZE_T)).
             I was confused while doing my previous change since
             fortran docs were wrong, and the type of the same argument in
             1_4 branch is INTEGER. I forgot to check carefully #if def for
             1_4 branch in the C code, and encountered only int type for the
             corresponding C functions. Since Fortran doesn't support
             backward compatibility, type of the argument should be
             always in sync with the latest branch.

Solution: ALWAYS, ALWAYS pay attention to compiler's warning.
          Fixed type to be INTEGER(SIZE_T).

Platforms tested: kelgia, copper 64-bit mode, solaris 2.7 64-bit mode.

Misc. update:

+++ fortran/src/H5Pff.f90
@@ -2110,7 +2110,7 @@
             INTEGER(HID_T), INTENT(IN) :: prp_id ! Property list identifier 
             INTEGER, INTENT(IN) :: mdc_nelmts  !Number of elements (objects)
                                                         ! in the meta data cache
-            INTEGER, INTENT(IN) :: rdcc_nelmts  !Number of elements (objects)
+            INTEGER(SIZE_T), INTENT(IN) :: rdcc_nelmts  !Number of elements (objects)
                                                         ! in the meta data cache
             INTEGER(SIZE_T), INTENT(IN) :: rdcc_nbytes !Total size of the raw data 
                                                       !chunk cache, in bytes 
@@ -2128,7 +2128,7 @@
               !DEC$ ENDIF
               INTEGER(HID_T), INTENT(IN) :: prp_id
               INTEGER, INTENT(IN) :: mdc_nelmts 
-              INTEGER, INTENT(IN) :: rdcc_nelmts 
+              INTEGER(SIZE_T), INTENT(IN) :: rdcc_nelmts 
               INTEGER(SIZE_T), INTENT(IN) :: rdcc_nbytes
               REAL, INTENT(IN) :: rdcc_w0
               END FUNCTION h5pset_cache_c
